fix: email verify include redirect
