+++ apps/api-users/src/app/modules/user/user.resolver.ts
@@ -10,7 +10,7 @@ import {
 import { GraphQLError } from 'graphql'
 
 import { User } from '.prisma/api-users-client'
-import { auth, impersonateUser } from '@core/nest/common/firebaseClient'
+import { impersonateUser } from '@core/nest/common/firebaseClient'
 import { CurrentIPAddress } from '@core/nest/decorators/CurrentIPAddress'
 import { CurrentUser } from '@core/nest/decorators/CurrentUser'
 import { CurrentUserId } from '@core/nest/decorators/CurrentUserId'
@@ -54,7 +54,10 @@ export class UserResolver {
     @CurrentUserId() userId: string,
     @Args('input') input?: MeInput
   ): Promise<User> {
-    return await this.findOrFetchUser(userId, input?.redirect ?? undefined)
+    return await this.userService.findOrFetchUser(
+      userId,
+      input?.redirect ?? undefined
+    )
   }
 
   @Query()
@@ -120,78 +123,7 @@ export class UserResolver {
     __typename: 'User'
     id: string
   }): Promise<User> {
-    return await this.findOrFetchUser(reference.id, undefined)
-  }
-
-  private async findOrFetchUser(
-    userId: string,
-    redirect: string | undefined = undefined
-  ): Promise<User> {
-    const existingUser = await this.prismaService.user.findUnique({
-      where: {
-        userId
-      }
-    })
-    if (existingUser != null && existingUser.emailVerified == null) {
-      const user = await this.prismaService.user.update({
-        where: {
-          id: userId
-        },
-        data: {
-          emailVerified: false
-        }
-      })
-      return user
-    }
-
-    if (existingUser != null && existingUser.emailVerified != null)
-      return existingUser
-
-    const {
-      displayName,
-      email,
-      emailVerified,
-      photoURL: imageUrl
-    } = await auth.getUser(userId)
-
-    const firstName = displayName?.split(' ')?.slice(0, -1)?.join(' ') ?? ''
-    const lastName = displayName?.split(' ')?.slice(-1)?.join(' ') ?? ''
-
-    const data = {
-      userId,
-      firstName,
-      lastName,
-      email: email ?? '',
-      imageUrl,
-      emailVerified
-    }
-
-    let user: User | null = null
-    let retry = 0
-
-    // this function can run in parallel as such it is possible for multiple
-    // calls to reach this point and try to create the same user
-    // due to the earlier firebase async call.
-    try {
-      user = await this.prismaService.user.create({
-        data
-      })
-      // after user create so it is ony sent once
-      if (!emailVerified && email != null) {
-        await this.userService.verifyUser(userId, email, redirect)
-      }
-    } catch (e) {
-      do {
-        user = await this.prismaService.user.update({
-          where: {
-            id: userId
-          },
-          data
-        })
-        retry++
-      } while (user == null && retry < 3)
-    }
-    return user
+    return await this.userService.findOrFetchUser(reference.id)
   }
 
   @Mutation()
