+++ apps/api-users/src/app/modules/user/user.service.ts
@@ -3,7 +3,8 @@ import { Injectable } from '@nestjs/common'
 import { Queue } from 'bullmq'
 import { getAuth } from 'firebase-admin/auth'
 
-import { firebaseClient } from '@core/nest/common/firebaseClient'
+import { User } from '.prisma/api-users-client'
+import { auth, firebaseClient } from '@core/nest/common/firebaseClient'
 
 import { PrismaService } from '../../lib/prisma.service'
 import { VerifyUserJob } from '../email/email.consumer'
@@ -15,6 +16,77 @@ export class UserService {
     private readonly emailQueue: Queue<VerifyUserJob>,
     private readonly prismaService: PrismaService
   ) {}
+
+  async findOrFetchUser(
+    userId: string,
+    redirect: string | undefined = undefined
+  ): Promise<User> {
+    const existingUser = await this.prismaService.user.findUnique({
+      where: {
+        userId
+      }
+    })
+    if (existingUser != null && existingUser.emailVerified == null) {
+      const user = await this.prismaService.user.update({
+        where: {
+          id: userId
+        },
+        data: {
+          emailVerified: false
+        }
+      })
+      return user
+    }
+
+    if (existingUser != null && existingUser.emailVerified != null)
+      return existingUser
+
+    const {
+      displayName,
+      email,
+      emailVerified,
+      photoURL: imageUrl
+    } = await auth.getUser(userId)
+
+    const firstName = displayName?.split(' ')?.slice(0, -1)?.join(' ') ?? ''
+    const lastName = displayName?.split(' ')?.slice(-1)?.join(' ') ?? ''
+
+    const data = {
+      userId,
+      firstName,
+      lastName,
+      email: email ?? '',
+      imageUrl,
+      emailVerified
+    }
+
+    let user: User | null = null
+    let retry = 0
+
+    // this function can run in parallel as such it is possible for multiple
+    // calls to reach this point and try to create the same user
+    // due to the earlier firebase async call.
+    try {
+      user = await this.prismaService.user.create({
+        data
+      })
+      // after user create so it is ony sent once
+      if (!emailVerified && email != null) {
+        await this.verifyUser(userId, email, redirect)
+      }
+    } catch (e) {
+      do {
+        user = await this.prismaService.user.update({
+          where: {
+            id: userId
+          },
+          data
+        })
+        retry++
+      } while (user == null && retry < 3)
+    }
+    return user
+  }
 
   async verifyUser(
     userId: string,
